test: refactor zip extraction
